created signing package under the domain package and moved the signing related classes there because it's better
