+++ app/src/main/java/com/babylon/wallet/android/domain/usecases/signing/SignTransactionUseCase.kt
@@ -1,10 +1,12 @@
-package com.babylon.wallet.android.domain.usecases
+package com.babylon.wallet.android.domain.usecases.signing
 
 import com.babylon.wallet.android.data.repository.transaction.TransactionRepository
 import com.babylon.wallet.android.data.transaction.NotaryAndSigners
 import com.babylon.wallet.android.data.transaction.TransactionConfig.EPOCH_WINDOW
 import com.babylon.wallet.android.domain.RadixWalletException
-import com.babylon.wallet.android.domain.model.SignRequest
+import com.babylon.wallet.android.domain.model.signing.SignRequest
+import com.babylon.wallet.android.domain.usecases.NotariseTransactionUseCase
+import com.babylon.wallet.android.domain.usecases.ResolveNotaryAndSignersUseCase
 import com.babylon.wallet.android.presentation.accessfactorsources.AccessFactorSourcesInput
 import com.babylon.wallet.android.presentation.accessfactorsources.AccessFactorSourcesProxy
 import com.radixdlt.sargon.AccountAddress
@@ -106,7 +108,7 @@ class SignTransactionUseCase @Inject constructor(
                         signRequest = signRequest
                     )
                 ).mapCatching { result ->
-                    result.signaturesWithPublicKey
+                    result.signersWithSignatures.values.toList()
                 }
             }
         }
