+++ app/src/main/java/com/babylon/wallet/android/domain/usecases/signing/SignWithLedgerFactorSourceUseCase.kt
@@ -1,10 +1,12 @@
-package com.babylon.wallet.android.domain.usecases.transaction
+package com.babylon.wallet.android.domain.usecases.signing
 
 import com.babylon.wallet.android.data.dapp.LedgerMessenger
 import com.babylon.wallet.android.data.dapp.model.LedgerDeviceModel
 import com.babylon.wallet.android.data.dapp.model.LedgerInteractionRequest
 import com.babylon.wallet.android.domain.model.IncomingMessage
-import com.babylon.wallet.android.domain.model.SignRequest
+import com.babylon.wallet.android.domain.model.IncomingMessage.LedgerResponse.SignatureOfSigner
+import com.babylon.wallet.android.domain.model.signing.EntityWithSignature
+import com.babylon.wallet.android.domain.model.signing.SignRequest
 import com.radixdlt.sargon.FactorSource
 import com.radixdlt.sargon.HierarchicalDeterministicPublicKey
 import com.radixdlt.sargon.PublicKey
